refactor fetch to use thunk
