+++ package.json
@@ -13,6 +13,7 @@
     "react-scripts": "5.0.0",
     "redux": "^4.1.2",
     "redux-logger": "^3.0.6",
+    "redux-thunk": "^2.4.1",
     "sass": "^1.49.9",
     "web-vitals": "^2.1.4"
   },
