+++ src/App.js
@@ -6,23 +6,12 @@ import NavBar from './components/NavBar';
 import MyProfile from './components/MyProfile';
 import RocketsPage from './components/RocketsPage';
 import Missions from './components/Missions';
-import { addRocket } from './redux/rockets/rockets';
+import { filterRocketInfo } from './redux/rockets/rockets';
 
 function App() {
-  const url = 'https://api.spacexdata.com/v3/rockets';
   const dispatch = useDispatch();
   useEffect(() => {
-    fetch(url)
-      .then((res) => res.json())
-      .then((data) => data.forEach((item) => {
-        const newRocket = {
-          id: item.id,
-          rocket_name: item.rocket_name,
-          description: item.description,
-          flickr_images: item.flickr_images,
-        };
-        dispatch(addRocket(newRocket));
-      }));
+    dispatch(filterRocketInfo());
   }, []);
   return (
     <div className="App">
