+++ src/redux/configStore.js
@@ -1,11 +1,12 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
+import thunk from 'redux-thunk';
 import rocketsReducer from './rockets/rockets';
 
 const reducer = combineReducers({
   rocketsReducer,
 });
 
-const store = createStore(reducer, applyMiddleware(logger));
+const store = createStore(reducer, applyMiddleware(logger, thunk));
 
 export default store;
